requirements.txt
beautifulsoup4==4.9.1
blis==0.4.1
Bottleneck==1.3.2
catalogue==1.0.0
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
cycler==0.10.0
cymem==2.0.3
dataclasses==0.7
fastai==1.0.61
fastprogress==0.2.3
Flask==1.1.2
Flask-Images==3.0.2
future==0.18.2
idna==2.10
importlib-metadata==1.7.0
itsdangerous==1.1.0
Jinja2==2.11.2
kiwisolver==1.2.0
MarkupSafe==1.1.1
matplotlib==3.2.2
murmurhash==1.0.2
numexpr==2.7.1
numpy==1.19.0
nvidia-ml-py3==7.352.0
packaging==20.4
pandas==1.0.5
PIL==1.1.7
Pillow==7.1.2
pillowcase==2.0.0
plac==1.1.3
preshed==3.0.2
pyparsing==2.4.7
python-dateutil==2.8.1
python-dotenv==0.13.0
pytz==2020.1
PyYAML==5.3.1
requests==2.24.0
scipy==1.5.0
six==1.15.0
soupsieve==2.0.1
spacy==2.3.0
srsly==1.0.2
thinc==7.4.1
torch==1.5.1
torchvision==0.6.1
tqdm==4.46.1
urllib3==1.25.9
wasabi==0.7.0
Werkzeug==1.0.1
zipp==3.1.0
